Inverse Reverse Cholesky

Compute inverse of a Cholesky matrix, storing the result in reverse order, see Details.

inverse_chol_reverse(Sigma_chol, lower_chol = FALSE)

Arguments

Sigma_chol
Holesky factor of a covariance matrix

lower_chol
TRUE if lower triangular Cholesky factors are used Default: FALSE

Value

A Cholesky matrix of the same triangle orientation as the input, with rows and columns in reverse order.

Details

For lower_chol == FALSE, the input Sigma_chol is the matrix \(R\) in the Cholesky factorisation \(\Sigma=R^T R\) with inverse \(Q=R^{-1}R^{-T}\). Since \(R^{-1}\) has the opposite upper/lower triangular property to \(R^T\), this \(Q\) factorisation isn't a regular Cholesky factorisation. Let \(P\) be the permutation matrix that reverses element order. Then \(PQP=PR^{-1}PPR^{-T}P\), and \(PR^{-T}P\) is the Cholesky factor of \(PQP\) with the same upper/lower triangular property as \(R\).

For lower_chol == TRUE, the input is \(L\) in \(\Sigma=LL^T\) and the output is \(PL^{-T}P\), with \(PQP=PL^{-T}PPL^{-1}P\).

Examples

if(interactive()){
  inverse_chol_reverse(matrix(c(1, 0, 2, 3), 2, 2))
}