Multiprobit Summary Information

Organise information about multiprobit parameter estimates and models

# S3 method for mp_estimate
summary(object, ...)

# S3 method for mp_model
summary(object, ...)

Arguments

object
An mp_estimate or mp_modelobject

...
Additional parameters, currently unused.

Value

An mp_estimate_summary or mp_model_summary list object with elements

beta

Summary information about the \(B\) coefficients (see mp_model() for model definition). The elements of the matrix \(B\) are listed row-wise, so that the values for each covariate are grouped together, for all Y-dimensions.

u

Summary information about the latent \(u\) parameters (see mp_model() for model definition). This is the internal scale representation of the probit correlation matrix parameter \(\Sigma\).

Sigma

Summary information for \(\Sigma\).

For model summaries, the prior mean and standard deviations are provided. For estimate summaries, the available quantities depend on what was computed and stored in the input object. Currently, the latent MAP estimates and their Hessian-based uncertainties are provided, and the corresponding point estimate of the probit correlation matrix parameter \(\Sigma\).

Examples

if (FALSE) {
if (interactive()) {
  # EXAMPLE1
}
}